chore: move contracts to data folder
